sse bug plus unit conversion tests
git-svn-id: file:///home/amuse/svn/amuse/trunk@90 6e832d89-e7d1-4ad1-8ca1-873beecccf65

+++ src/amuse/legacy/sse/muse_worker.cc
@@ -168,7 +168,7 @@ void run_loop() {
 			parent.Send(doubles_out, reply_header.number_of_doubles, MPI_DOUBLE, 0, 999);
 		}
 		if( reply_header.number_of_ints > 0) {
-			parent.Send(ints_out, reply_header.number_of_ints, MPI_DOUBLE, 0, 999);
+			parent.Send(ints_out, reply_header.number_of_ints, MPI_INT, 0, 999);
 		}
 	}
